fix: Use int64 for ids and latency

+++ endpoint/service.go
@@ -7,7 +7,7 @@ import (
 type Storage interface {
 	InsertEndpoint(ctx context.Context, item Endpoint) error
 	GetAllEndpoints(ctx context.Context) ([]Endpoint, error)
-	GetEndpoint(ctx context.Context, id int) (Endpoint, error)
+	GetEndpoint(ctx context.Context, id int64) (Endpoint, error)
 	UpdateEndpoint(ctx context.Context, item Endpoint) error
 }
 
@@ -29,7 +29,7 @@ func (s *EndpointService) GetAll(ctx context.Context) ([]Endpoint, error) {
 	return s.storage.GetAllEndpoints(ctx)
 }
 
-func (s *EndpointService) Get(ctx context.Context, id int) (Endpoint, error) {
+func (s *EndpointService) Get(ctx context.Context, id int64) (Endpoint, error) {
 	return s.storage.GetEndpoint(ctx, id)
 }
